move my cloud stuff to pavels branchswitch thing
